Changed to double again
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@18727 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: 9f4e68b99c5af4b6152986f94adac0a3b8bcd6fb

+++ server/relational/relational_interval/src/de/uniol/inf/is/odysseus/relational_interval/physicaloperator/RelationalFastMedianHistogramPO.java
@@ -12,7 +12,7 @@ public class RelationalFastMedianHistogramPO<T extends Comparable<T>>
 		AbstractFastMedianPO<T> {
 
 	Map<Long, Histogram<T, Tuple<? extends ITimeInterval>>> elements = new HashMap<>();
-	long roundfactor = 100;
+	double roundfactor = 100;
 
 	public RelationalFastMedianHistogramPO(int medianAttrPos, boolean numericalMedian, long roundfactor) {
 		super(medianAttrPos , numericalMedian);
@@ -29,7 +29,7 @@ public class RelationalFastMedianHistogramPO<T extends Comparable<T>>
 				groupID, medianAttribute);
 		groupHistogram.cleanUp(object.getMetadata().getStart());
 		if (roundfactor>0){
-			groupHistogram.addElement((T)new Double((Math.round(((Number)medianAttribute).doubleValue()*roundfactor)/roundfactor)), object);
+			groupHistogram.addElement((T)new Double((Math.round(((Number)medianAttribute).doubleValue()*roundfactor)/(roundfactor))), object);
 		}else{
 			groupHistogram.addElement(medianAttribute, object);
 		}
